Extract helpers functions to e2e-test-utils
