+++ packages/e2e-test-utils/index.ts
@@ -6,3 +6,4 @@
  * CERTAIN THIRD PARTY SOFTWARE MAY BE CONTAINED IN PORTIONS OF THE SOFTWARE. See NOTICE FILE AT https://github.com/salto-io/salto/blob/main/NOTICES
  */
 export * from './src/e2e_utils'
+export * as helpers from './src/helpers'
