+++ packages/e2e-test-utils/src/helpers.ts
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2025 Salto Labs Ltd.
+ * Licensed under the Salto Terms of Use (the "License");
+ * You may not use this file except in compliance with the License.  You may obtain a copy of the License at https://www.salto.io/terms-of-use
+ *
+ * CERTAIN THIRD PARTY SOFTWARE MAY BE CONTAINED IN PORTIONS OF THE SOFTWARE. See NOTICE FILE AT https://github.com/salto-io/salto/blob/main/NOTICES
+ */
+
+import { DetailedChangeWithBaseChange, InstanceElement, toChange } from '@salto-io/adapter-api'
+import { getDetailedChanges } from '@salto-io/adapter-utils'
+
+export const getAdditionDetailedChangesFromInstances = (
+  instances: InstanceElement[],
+): DetailedChangeWithBaseChange[] => {
+  const changes = instances.map(inst => toChange({ after: inst }))
+  return changes.flatMap(change => getDetailedChanges(change))
+}
+
+export const getDeletionDetailedChangesFromInstances = (
+  instances: InstanceElement[],
+): DetailedChangeWithBaseChange[] => {
+  const changes = instances.map(inst => toChange({ before: inst }))
+  return changes.flatMap(change => getDetailedChanges(change))
+}
